[refactor] 백엔드 Global 패키지 리팩토링 (#174)
* refactor: Diary에서 쓰는 Converter

* test 변경

* refactor: BaseEntity 상속구조 제거

* refactor: HealthController 패키지 변경

* refactor: ImageConverter 책임을  DIaryS3Service로 이동

* refactor: XmlMessageSource 패키지 변경

* feat: 헬스 컨트롤러 제거 및 인프라 헬스 체크 엔드포인트 변경

* fix: @EntityListeners(AuditingEntityListener.class) 추가
