+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/domain/converter/DiaryLongListToStringConverter.java
@@ -1,6 +1,6 @@
-package com.startingblue.fourtooncookie.global.converter;
+package com.startingblue.fourtooncookie.diary.domain.converter;
 
-import com.startingblue.fourtooncookie.global.converter.exception.ConversionException;
+import com.startingblue.fourtooncookie.diary.exception.DiaryConversionException;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import org.springframework.stereotype.Component;
@@ -11,7 +11,7 @@ import java.util.stream.Collectors;
 
 @Converter
 @Component
-public class LongListToStringConverter implements AttributeConverter<List<Long>, String> {
+public class DiaryLongListToStringConverter implements AttributeConverter<List<Long>, String> {
 
     private static final String EMPTY_LIST_STRING = "";
 
@@ -26,7 +26,7 @@ public class LongListToStringConverter implements AttributeConverter<List<Long>,
                     .map(String::valueOf)
                     .collect(Collectors.joining(","));
         } catch (Exception e) {
-            throw new ConversionException("Error converting List<Long> to String", e);
+            throw new DiaryConversionException("Error converting List<Long> to String", e);
         }
     }
 
@@ -41,7 +41,7 @@ public class LongListToStringConverter implements AttributeConverter<List<Long>,
                     .map(Long::valueOf)
                     .toList();
         } catch (NumberFormatException e) {
-            throw new ConversionException("Error converting String to List<Long>", e);
+            throw new DiaryConversionException("Error converting String to List<Long>", e);
         }
     }
