+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/domain/converter/DiaryPaintingImageGenerationStatusListConverter.java
@@ -1,7 +1,7 @@
 package com.startingblue.fourtooncookie.diary.domain.converter;
 
 import com.startingblue.fourtooncookie.diary.domain.DiaryPaintingImageGenerationStatus;
-import com.startingblue.fourtooncookie.global.converter.exception.ConversionException;
+import com.startingblue.fourtooncookie.diary.exception.DiaryConversionException;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import org.springframework.stereotype.Component;
@@ -28,7 +28,7 @@ public class DiaryPaintingImageGenerationStatusListConverter implements Attribut
                     .map(DiaryPaintingImageGenerationStatus::name)
                     .collect(Collectors.joining(","));
         } catch (Exception e) {
-            throw new ConversionException("Error converting List<DiaryImageGenerationStatus> to String", e);
+            throw new DiaryConversionException("Error converting List<DiaryImageGenerationStatus> to String", e);
         }
     }
 
@@ -43,7 +43,7 @@ public class DiaryPaintingImageGenerationStatusListConverter implements Attribut
                     .map(DiaryPaintingImageGenerationStatus::valueOf)
                     .toList();
         } catch (IllegalArgumentException e) {
-            throw new ConversionException("Error converting String to List<DiaryImageGenerationStatus>", e);
+            throw new DiaryConversionException("Error converting String to List<DiaryImageGenerationStatus>", e);
         }
     }
