+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/domain/converter/DiaryUrlListToStringConverter.java
@@ -1,6 +1,6 @@
-package com.startingblue.fourtooncookie.global.converter;
+package com.startingblue.fourtooncookie.diary.domain.converter;
 
-import com.startingblue.fourtooncookie.global.converter.exception.ConversionException;
+import com.startingblue.fourtooncookie.diary.exception.DiaryConversionException;
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
 import org.springframework.stereotype.Component;
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 
 @Converter
 @Component
-public class UrlListToStringConverter implements AttributeConverter<List<URL>, String> {
+public class DiaryUrlListToStringConverter implements AttributeConverter<List<URL>, String> {
 
     private static final String EMPTY_LIST_STRING = "";
 
@@ -28,7 +28,7 @@ public class UrlListToStringConverter implements AttributeConverter<List<URL>, S
                     .map(String::valueOf)
                     .collect(Collectors.joining(","));
         } catch (Exception e) {
-            throw new ConversionException("Error converting List<URL> to String", e);
+            throw new DiaryConversionException("Error converting List<URL> to String", e);
         }
     }
 
@@ -44,12 +44,12 @@ public class UrlListToStringConverter implements AttributeConverter<List<URL>, S
                         try {
                             return new URL(urlString);
                         } catch (MalformedURLException e) {
-                            throw new ConversionException("Malformed URL: " + urlString, e);
+                            throw new DiaryConversionException("Malformed URL: " + urlString, e);
                         }
                     })
                     .toList();
         } catch (Exception e) {
-            throw new ConversionException("Error converting String to List<URL>", e);
+            throw new DiaryConversionException("Error converting String to List<URL>", e);
         }
     }
