+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/exception/DiaryConversionException.java
@@ -1,11 +1,11 @@
-package com.startingblue.fourtooncookie.global.converter.exception;
+package com.startingblue.fourtooncookie.diary.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-public class ConversionException extends RuntimeException {
-    public ConversionException(String message, Throwable cause) {
+public class DiaryConversionException extends RuntimeException {
+    public DiaryConversionException(String message, Throwable cause) {
         super(message, cause);
     }
 }
