--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/global/monitoring/HealthController.java
@@ -1,16 +1,0 @@
-package com.startingblue.fourtooncookie.global.monitoring;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/health")
-public class HealthController {
-
-    @GetMapping
-    public HttpStatus health() {
-        return HttpStatus.OK;
-    }
-}
