+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/domain/Member.java
@@ -1,9 +1,11 @@
 package com.startingblue.fourtooncookie.member.domain;
 
-import com.startingblue.fourtooncookie.global.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -15,7 +17,8 @@ import java.util.UUID;
 @Builder
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member extends BaseEntity {
+@EntityListeners(AuditingEntityListener.class)
+public class Member {
 
     @Id
     @Column(name = "member_id")
@@ -34,6 +37,12 @@ public class Member extends BaseEntity {
     @Enumerated(EnumType.STRING)
     @Column(name = "role")
     private Role role;
+
+    @CreatedDate
+    private LocalDateTime createdDateTime;
+
+    @LastModifiedDate
+    private LocalDateTime modifiedDateTime;
 
     @Column(name = "deleted_date_time")
     private LocalDateTime deletedDateTime;
