+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/character/service/CharacterServiceTest.java
@@ -1,7 +1,7 @@
 package com.startingblue.fourtooncookie.character.service;
 
+import com.startingblue.fourtooncookie.artwork.ArtworkService;
 import com.startingblue.fourtooncookie.artwork.domain.Artwork;
-import com.startingblue.fourtooncookie.artwork.service.ArtworkService;
 import com.startingblue.fourtooncookie.character.CharacterService;
 import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.character.CharacterRepository;
